requirements.txt
aiohttp>=3.10.2
langchain-core>=0.2.11
langchain>=0.2.11
openai>=1.57.0
pandas
groq>=0.11.0
pypdf>=5.3.1
google-genai>=1.3.0
Markdown>=3.7
litellm>=1.51.1
tenacity==8.3.0
tqdm>=4.66.5
llama-index>=0.10.0
pyopenssl>=24.2.1
psutil~=6.0.0
py-cpuinfo~=9.0.0
requests~=2.32.3
GPUtil~=1.4.0
ipynbname
tiktoken>=0.7.0
tomli>=2.0.0
rich>=13.9.4
openinference-instrumentation-llama-index
openinference-instrumentation-langchain
openinference-instrumentation-vertexai
openinference-instrumentation-anthropic
openinference-instrumentation-groq
openinference-instrumentation-litellm
openinference-instrumentation-mistralai
openinference-instrumentation-openai
openinference-instrumentation-bedrock
openinference-instrumentation-crewai
openinference-instrumentation-haystack
openinference-instrumentation-openai-agents
openinference-instrumentation-smolagents
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-proto>=1.12.0